+++ src/main/java/com/ye/vio/service/impl/TopicLikeServiceImpl.java
@@ -1,13 +1,19 @@
 package com.ye.vio.service.impl;
 
+import com.ye.vio.dao.NotificationLikeDao;
 import com.ye.vio.dao.TopicLikeDao;
+import com.ye.vio.entity.NotificationLike;
 import com.ye.vio.entity.TopicLike;
+import com.ye.vio.enums.CustomizeErrorCode;
+import com.ye.vio.exception.CustomizeException;
+import com.ye.vio.service.NotificationLikeService;
 import com.ye.vio.util.UUIDUtils;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.annotation.Resource;
 import java.util.Date;
+import java.util.List;
 
 /**
  * @program: vio
@@ -21,25 +27,46 @@ public class TopicLikeServiceImpl implements com.ye.vio.service.TopicLikeService
 
     @Resource
     private TopicLikeDao topicLikeDao;
+    @Resource
+    private NotificationLikeDao notificationLikeDao;
+
 
 
     @Override
     @Transactional
     public int addTopicLike(TopicLike topicLike) {
+
+
+        if(topicLike.getUserId()==null||topicLike.getToUserId()==null||topicLike.getLikedTopicId()==null)
+            throw new CustomizeException(CustomizeErrorCode.INVALID_INPUT);
+       List<TopicLike> topicLikeList= topicLikeDao.queryTopicLikeByTopicIdAndUserId(topicLike);
+        if(topicLikeList!=null&&topicLikeList.size()>0)
+            throw new  CustomizeException(CustomizeErrorCode.TOPIC_LIKE_REPEAT_ERROR);
+
         topicLike.setTopicLikeId(UUIDUtils.UUID());
         topicLike.setCreateTime(new Date());
 
-        int effected=0;
+        int effected=topicLikeDao.insertTopicLike(topicLike);
 
-            effected=topicLikeDao.insertTopicLike(topicLike);
-
-            if(effected<=0)throw new RuntimeException("添加话题点赞失败");
+            if(effected<=0)throw new CustomizeException(CustomizeErrorCode.TOPIC_LIKE_ADD_ERROR);
 
             int effectedUpdate=topicLikeDao.updateTopicLikeNum(1,topicLike.getLikedTopicId());
 
-            if (effectedUpdate<=0) throw new RuntimeException("更新话题点赞数失败");
+            if (effectedUpdate<=0) throw new CustomizeException(CustomizeErrorCode.TOPIC_LIKE_UPDATE_REPLY_ERROR);
 
+        NotificationLike notificationLike=new NotificationLike();
 
+        notificationLike.setNotificationLikeId(UUIDUtils.UUID());//设置唯一标识
+        notificationLike.setTopicId(topicLike.getLikedTopicId());//点赞话题id
+        notificationLike.setFromUserId(topicLike.getUserId());//谁点的赞
+        notificationLike.setToUserId(topicLike.getToUserId());//给谁
+        notificationLike.setNotificationType(1);//什么类型，话题
+        notificationLike.setCreateTime(new Date());
+
+       int effectedNoti= notificationLikeDao.insertNotificationLike(notificationLike);
+
+       if(effectedNoti<=0)
+           throw new CustomizeException(CustomizeErrorCode.NOTI_LIKE_ADD_ERROR);
 
         return effected;
     }
@@ -48,15 +75,16 @@ public class TopicLikeServiceImpl implements com.ye.vio.service.TopicLikeService
     @Override
     @Transactional
     public int removeTopicLike(TopicLike topicLike) {
-        int effectedDelete=0;
+        if(topicLike.getLikedTopicId()==null||topicLike.getUserId()==null)
+            throw new CustomizeException(CustomizeErrorCode.INVALID_INPUT);
 
-            effectedDelete=topicLikeDao.deleteTopicLike(topicLike.getUserId(),topicLike.getLikedTopicId());
+        int effectedDelete=topicLikeDao.deleteTopicLike(topicLike.getUserId(),topicLike.getLikedTopicId());
 
-            if(effectedDelete<=0)throw new RuntimeException("删除话题点赞失败");
+            if(effectedDelete<=0)throw new CustomizeException(CustomizeErrorCode.TOPIC_LIKE_DELETE_ERROR);
 
             int effectedUpdate=topicLikeDao.updateTopicLikeNum(2,topicLike.getLikedTopicId());
 
-            if (effectedUpdate<=0) throw new RuntimeException("更新话题点赞数失败");
+            if (effectedUpdate<=0) new CustomizeException(CustomizeErrorCode.TOPIC_LIKE_UPDATE_REPLY_ERROR);
